Add files via upload
An updated version of the dilution series script that creates two worklists (.gwl-files) for importing into EvoWare, one for the first half which uses 1000ul tips and one for the second half which uses 200ul tips. This just means that when you want to run a script like this on the robot you have to import both the 1000ul worklist and the 200ul worklist, selecting to use 1000ul tips for the first one and 200ul tips for the second one.